requirements.txt
streamlit
streamlit_option_menu
streamlit-navigation-bar
python_dotenv
langchain-ollama
langchain-community
PyPDF2
pypdf
langchain
langchain-text-splitters
chromadb
faiss-cpu
audio-recorder-streamlit
langchain-groq
langchain-huggingface
elevenlabs
groq
sentence-transformers